requirements.txt

python-multipart
pillow
tensorflow-serving-api==2.5.0
matplotlib
numpy